refactor: apply pr suggestions
